+++ Changelog.md
@@ -7,16 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 7.78.0
+
 ### Added
 - This changelog.
 - Open source button to README.md.
 - GCode headers now include M201, M203, M204, and M205 for speed tuning.
 - English for "Jerk" in plotter settings tab.
+- Added Peano and other L system fractals, all consolidated under `Fractals > L Systems`.
 ### Changed
 - Size of "upload firmware" dialog should now be large enough to see all text.
 - Bumped Jogamp to 2.6.0.
 ### Deprecated
 ### Removed
+- use of OpenGL and all hardware accelerated graphics
 ### Fixed
 - English label for "Acceleration" in plotter settings tab.
 - Calculation for meters of timing belt needed on each side of Makelangelo 5 and HUGE.